# package/x12/x12/api.yml
openapi: 3.0.3
info:
  version: "0.0.0"
  description: "X12 EDI file receiver (DataProducer)"
  title: '@zerobias-org/module-x12-x12'
  x-impl-name: x12
  x-product-infos:
    - $ref: './node_modules/@zerobias-org/product-auditmation-generic-dataproducer/catalog.yml#/Product'
    - $ref: './node_modules/@zerobias-org/product-x12-x12/catalog.yml#/Product'
# This module implements the DataProducer interface defined in
# @zerobias-org/module-interface-dataproducer. Interface paths are referenced
# directly from the interface module; the implementation is Java (see java/).
# Adds /objects/{objectId}/download (binary) over hl7/v2: each X12 file node is
# also a binary object whose raw EDI bytes are downloadable.
paths:
...
  /connect:
    post:
      tags: [connection]
      summary: Establish a receiver session
      description: >-
        Validates the connection profile and binds the producer session. The
        inbox poller and buffer are daemon-owned and already running; connect
        is the DataProducer handshake the pipeline uses before browse/drain.
      operationId: connect
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectionProfile'
      responses:
        '200':
          description: Connected
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: connected
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Connection failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ConnectionProfile:
      $ref: './connectionProfile.yml'
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Error message
        details:
          type: object
          additionalProperties: true
          description: Additional error details
...